+++ Oxygen/oxygenengine/source/oxygen/application/overlays/SaveStateMenu.cpp
@@ -21,7 +21,7 @@
 SaveStateMenu::SaveStateMenu()
 {
 	mFont.setSize(18.0f);
-	mFont.addFontProcessor(std::make_shared<ShadowFontProcessor>(Vec2f(2.0f, 2.0f), 0.5f));
+	mFont.addFontProcessor(std::make_shared<ShadowFontProcessor>(Vec2i(2, 2), 0.5f));
 }
 
 SaveStateMenu::~SaveStateMenu()
